refactor: move fetch logic to org context
